+++ src/ALU.cpp
@@ -1,9 +1,9 @@
 #include "ALU.h"
 #include <cstdint>
 
-uint16_t ALU::ADD(uint16_t a, uint16_t b){
+int16_t ALU::ADD(int16_t a, int16_t b){
     return a + b;
 }
-uint16_t ALU::MUL(uint16_t a, uint16_t b){
-    return a * b;
+int16_t ALU::MUL(int16_t a, int16_t b){
+    return int32_t(a) * int32_t(b);
 }
